# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "승뮤니티 API 문서화"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/login": {
      "get": {
        "tags": ["User"],
        "summary": "로그인 페이지 조회",
        "description": "로그인 페이지를 조회합니다.",
        "responses": {
          "200": {
            "description": "로그인 페이지 조회 성공"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "사용자 로그인",
        "description": "사용자를 인증하고 세션을 생성하여 로그인합니다.",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "사용자의 아이디"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "사용자의 비밀번호"
          }
        ],
        "responses": {
          "200": {
            "description": "로그인 성공"
          },
          "401": {
            "description": "잘못된 사용자 아이디 또는 비밀번호",
            "schema": {
              "type": "string",
              "example": "<script>alert('인증 정보가 유효하지 않습니다.');history.go(-1)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
...
    "/list": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 1페이지 조회",
        "description": "게시글의 1페이지를 조회합니다.",
        "responses": {
          "200": {
            "description": "게시글 1페이지 조회 성공"
          }
        }
      },
      "delete": {
        "tags": ["Post"],
        "summary": "특정 게시글 삭제",
        "description": "ID 값을 사용하여 특정 게시글을 삭제합니다.",
        "parameters": [
          {
            "name": "docid",
            "in": "query",
            "description": "게시물의 ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 삭제 완료"
          },
          "401": {
            "description": "사용자가 로그인하지 않음"
          },
          "403": {
            "description": "게시글을 삭제할 권한이 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/list/{num}": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 페이지별 조회",
        "description": "게시글을 페이지별로 조회합니다.",
        "parameters": [
          {
            "name": "num",
            "in": "path",
            "description": "페이지 번호",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 페이지 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "글목록": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Post"
                  }
                },
                "글수": {
                  "type": "integer",
                  "description": "전체 게시글 수"
                },
                "채팅사람": {
                  "type": "string",
                  "description": "채팅 상대자 이름"
                },
                "페이지넘버": {
                  "type": "integer",
                  "description": "현재 페이지 번호"
                },
                "resizeImg": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "description": "리사이즈된 이미지 URL"
                  }
                },
                "isRead": {
                  "type": "boolean",
                  "description": "읽음 여부"
                }
              }
            }
          },
          "404": {
            "description": "게시글이 존재하지 않음",
            "schema": {
              "type": "string",
              "example": "<script>alert('게시글이 존재하지 않습니다.');window.location.replace(`/list/1`)</script>"
            }
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
    "/list/next/{num}": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 다음 페이지 조회",
        "description": "게시글의 다음 페이지를 조회합니다.",
        "parameters": [
          {
            "name": "num",
            "in": "path",
            "description": "이전 페이지의 마지막 글의 ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "현재 페이지 번호",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "페이지 이동 성공",
            "schema": {
              "type": "object",
              "properties": {
                "글목록": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Post"
                  }
                },
                "글수": {
                  "type": "integer",
                  "description": "전체 글 수"
                },
                "채팅사람": {
                  "type": "string",
                  "description": "채팅 사용자명"
                },
                "페이지넘버": {
                  "type": "integer",
                  "description": "현재 페이지 번호"
                },
                "resizeImg": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "이미지 경로"
                },
                "isRead": {
                  "type": "boolean",
                  "description": "읽음 여부"
                }
              }
            }
          },
          "404": {
            "description": "다음 페이지가 존재하지 않음",
            "schema": {
              "type": "string",
              "example": "<script>alert('다음페이지가 존재하지 않습니다.');window.location.replace(`/list/1`)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/list/prev/{num}": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 이전 페이지 조회",
        "description": "게시글의 이전 페이지를 조회합니다.",
        "parameters": [
          {
            "name": "num",
            "in": "path",
            "description": "다음 페이지의 첫 번째 글 ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "현재 페이지 번호",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "페이지 이동 성공",
            "schema": {
              "type": "object",
              "properties": {
                "글목록": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Post"
                  }
                },
                "글수": {
                  "type": "integer",
                  "description": "전체 글 수"
                },
                "채팅사람": {
                  "type": "string",
                  "description": "채팅 사용자명"
                },
                "페이지넘버": {
                  "type": "integer",
                  "description": "현재 페이지 번호"
                },
                "resizeImg": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "이미지 경로"
                },
                "isRead": {
                  "type": "boolean",
                  "description": "읽음 여부"
                }
              }
            }
          },
          "404": {
            "description": "이전 페이지가 존재하지 않음",
            "schema": {
              "type": "string",
              "example": "<script>alert('이전페이지가 존재하지 않습니다.');window.location.replace(`/list/1`)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/list/search": {
      "get": {
        "tags": ["Post"],
        "summary": "특정 게시글 검색",
        "description": "특정 게시글을 검색합니다.",
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "description": "검색어",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 검색 성공",
            "schema": {
              "type": "object",
              "properties": {
                "글목록": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Post"
                  }
                },
                "resizeImg": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "이미지 경로"
                }
              }
            }
          },
          "400": {
            "description": "빈칸을 검색했을 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('검색어를 입력해주세요.');window.location.replace(`/list/1`)</script>"
            }
          },
          "404": {
            "description": "검색한 게시글이 존재하지 않는 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('존재하지 않는 글');window.location.replace(`/list/1`)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
...
    "/edit": {
      "put": {
        "tags": ["Post"],
        "summary": "게시글 수정",
        "description": "사용자가 게시글을 수정합니다.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "수정할 게시물 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "수정할 게시물의 ID"
                },
                "userId": {
                  "type": "string",
                  "description": "게시물을 작성한 사용자의 ID"
                },
                "title": {
                  "type": "string",
                  "description": "게시물 제목 (20글자 이하)"
                },
                "content": {
                  "type": "string",
                  "description": "게시물 내용"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 수정 성공"
          },
          "400": {
            "description": "게시글 제목이 20글자를 초과하는 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('제목을 20글자 이하로 지정해주세요.');history.go(-1)</script>"
            }
          },
          "403": {
            "description": "게시글을 수정할 권한이 없음",
            "schema": {
              "type": "string",
              "example": "<script>alert('수정할 수 없습니다.');window.location.replace('/list/1')</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/edit/{id}": {
      "get": {
        "tags": ["Post"],
        "summary": "게시물 수정 페이지 조회",
        "description": "특정 게시물을 수정하는 페이지를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "수정할 게시글의 ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "게시물 수정 페이지 조회 성공"
          },
          "404": {
            "description": "게시물을 찾을 수 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
...
  },
...
}